immunifhir-backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
click==8.1.8
cryptography==46.0.3
deprecation==2.1.0
exceptiongroup==1.3.0
fastapi==0.119.0
fhir.resources==8.1.0
fhir_core==1.1.4
h11==0.16.0
hl7apy==1.3.5
idna==3.11
multidict==6.7.0
packaging==25.0
postgrest==2.22.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.1.1
realtime==2.22.0
sniffio==1.3.1
starlette==0.48.0
storage3==2.22.0
StrEnum==0.4.15
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
websockets==15.0.1
xmltodict==1.0.2
yarl==1.22.0
